refactor: schema 폴더명 schemas로 통일
