+++ src/pages/@owner/estimate/hooks/use-estimate-date.ts
@@ -1,7 +1,7 @@
 import { type UseFormReturn } from 'react-hook-form';
 
 import type { AvailableDate } from '@type/available-date';
-import type { EstimateFormData } from '@pages/@owner/estimate/schema/estimate.schema';
+import type { EstimateFormData } from '@pages/@owner/estimate/schemas/estimate.schema';
 
 export const useEstimateDate = (methods: UseFormReturn<EstimateFormData>) => {
   const {
